Move classes in surveys package into surveys/list package
