+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.surveys.SurveysActivity">
+        <activity android:name=".ui.surveys.list.SurveysActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
