+++ app/src/main/java/com/oakkub/survey/ui/surveys/list/SurveysLoadingRequest.kt
@@ -1,4 +1,4 @@
-package com.oakkub.survey.ui.surveys
+package com.oakkub.survey.ui.surveys.list
 
 /**
  * Created by oakkub on 24/3/2018 AD.
