+++ app/src/main/java/com/oakkub/survey/ui/surveys/list/SurveysUiModel.kt
@@ -1,4 +1,4 @@
-package com.oakkub.survey.ui.surveys
+package com.oakkub.survey.ui.surveys.list
 
 import com.oakkub.survey.data.services.surveys.SurveyResponse
