+++ app/src/main/java/com/oakkub/survey/ui/surveys/list/adapter/content/SurveysItemAdapterModel.kt
