+++ app/src/main/res/layout/fragment_surveys.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.oakkub.survey.ui.surveys.SurveysFragment">
+    tools:context="com.oakkub.survey.ui.surveys.list.SurveysFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/surveysItemRecyclerView"
